ebs_stale_snapshosts.py

+++ Identifying-Stale-EBS-Snapshots/ebs_stale_snapshosts.py
@@ -2,27 +2,25 @@ import boto3
 
 def lambda_handler(event, context):
     ec2 = boto3.client('ec2')
+    
+    # Get the IAM user's AWS account ID
+    account_id = boto3.client('sts').get_caller_identity().get('Account')
 
-    # Get all EBS snapshots
-    response = ec2.describe_snapshots(OwnerIds=['self'])
+    # Get all EBS Snapshots owned by the user
+    response = ec2.describe_snapshots(OwnerIds=[account_id])
 
-    # Get all active EC2 instance IDs
-    instances_response = ec2.describe_instances(Filters=[{'Name': 'instance-state-name', 'Values': ['running']}])
-    active_instance_ids = set()
-
-    for reservation in instances_response['Reservations']:
-        for instance in reservation['Instances']:
-            active_instance_ids.add(instance['InstanceId'])
-
-    # Iterate through each snapshot and delete if it's not attached to any volume or the volume is not attached to a running instance
+    # Iterate through each snapshot and delete if it's not attached to any volume and not associated with any instance
     for snapshot in response['Snapshots']:
         snapshot_id = snapshot['SnapshotId']
         volume_id = snapshot.get('VolumeId')
+        
+        # Get the list of attachments associated with the snapshot
+        attachments = snapshot.get('Attachments', [])
 
-        if not volume_id:
-            # Delete the snapshot if it's not attached to any volume
+        # Delete the snapshot only if both volume_id and attachments are empty
+        if not volume_id and not attachments:
             ec2.delete_snapshot(SnapshotId=snapshot_id)
-            print(f"Deleted EBS snapshot {snapshot_id} as it was not attached to any volume.")
+            print(f"Deleted EBS snapshot {snapshot_id} as it was not attached to any volume and not associated with any instance.")
         else:
             # Check if the volume still exists
             try:
